--- src/test/java/blackjack/domain/CardTypes.java
@@ -1,19 +1,0 @@
-package blackjack.domain;
-
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Rank;
-import blackjack.domain.card.Suit;
-import java.util.ArrayList;
-import java.util.List;
-
-public class CardTypes {
-    public static List<Card> createDeck() {
-        List<Card> deck = new ArrayList<>();
-        for (Suit suit : Suit.values()) {
-            for (Rank rank : Rank.values()) {
-                deck.add(new Card(suit, rank));
-            }
-        }
-        return deck;
-    }
-}
