+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,9 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Deck;
-import blackjack.domain.card.Rank;
-import blackjack.domain.card.Suit;
 import java.util.List;
 import java.util.Stack;
 import org.assertj.core.api.Assertions;
