+++ src/test/java/blackjack/domain/card/ScoreCalculatorTest.java
@@ -1,9 +1,5 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.card.Rank;
-import blackjack.domain.card.ScoreCalculator;
-import blackjack.domain.card.Suit;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
