+++ src/test/java/blackjack/domain/participants/DealerTest.java
@@ -1,14 +1,12 @@
-package blackjack.domain;
+package blackjack.domain.participants;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import blackjack.domain.card.Deck;
+import blackjack.domain.card.FixCardsShuffler;
 import blackjack.domain.card.Rank;
 import blackjack.domain.card.ScoreCalculator;
 import blackjack.domain.card.Suit;
-import blackjack.domain.participants.Dealer;
-import blackjack.domain.participants.Player;
-import blackjack.domain.participants.Players;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Stack;
